backend/pyproject.toml
[project]
name = "booms-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "openai (>=2.15.0,<3.0.0)",
    "weasyprint (>=68.0,<69.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.14.14,<0.15.0)"
]
